+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { RouterProvider } from "react-router-dom";
-import { appRouter } from "./appRouter";
+import { appRouter } from "./routes/appRouter";
 
 function App() {
 	return <RouterProvider router={appRouter} />;
