+++ src/layouts/DefaultLayout.tsx
@@ -1,0 +1,17 @@
+import { ReactNode } from "react";
+import { Header } from "../components/Header/Header";
+import { Footer } from "../components/Footer/Footer";
+
+type LayoutProps = {
+	children: ReactNode;
+};
+
+export const DefaultLayout: React.FC<LayoutProps> = ({ children }) => {
+	return (
+		<>
+			<Header />
+			{children}
+			<Footer />
+		</>
+	);
+};
